fix: add perm check

+++ src/main/java/de/jaunikapauni/axclans/command/DeleteCommand.java
@@ -15,6 +15,10 @@ public class DeleteCommand implements CommandExecutor {
     @Override
     public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, @NotNull String @NotNull [] args) {
         Player p = (Player) sender;
+        if(!p.hasPermission("axclans.create")){
+            p.sendMessage("You don't have the permission! [axclans.create]");
+            return true;
+        }
         String name = args[0];
         reference.getClanManager().deleteClan(name);
         p.sendMessage("You successfully deleted the clan " + name);
